requirements.txt
pandas
flask_sqlalchemy
numpy
scikit-learn
plotly